requirements.txt
amqp==5.3.1
asgiref==3.8.1
awsebcli==3.21.0
billiard==4.2.1
boto3==1.35.79
botocore==1.35.79
celery==5.5.3
cement==2.10.14
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.0
distlib==0.3.9
dj-database-url==3.0.1
Django==5.1.2
django-cors-headers==4.7.0
django-crontab==0.7.1
django-environ==0.12.0
django-filter==24.3
django-ratelimit==4.1.0
django-redis==6.0.0
django-storages==1.14.4
django-taggit==6.1.0
django_csp==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
drf-yasg==1.21.10
filelock==3.16.1
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
kombu==5.5.4
mysql-connector-python==9.0.0
mysqlclient==2.2.5
packaging==24.2
parameterized==0.9.0
pathspec==0.10.1
pillow==11.0.0
platformdirs==4.3.6
promise==2.3
prompt_toolkit==3.0.52
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.2
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==308
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
s3transfer==0.10.4
semantic-version==2.10.0
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.1
termcolor==2.5.0
text-unidecode==1.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.20
vine==5.1.0
virtualenv==20.28.0
wcwidth==0.2.13
